Refactor ibc directory
