+++ .github/workflows/golangci-lint.yml
@@ -30,7 +30,7 @@ jobs:
           # working-directory: somedir
 
           # Optional: golangci-lint command line arguments.
-          args: --timeout=20m
+          args: --timeout=20m --skip-dirs legacy_ibc_testing
 
           # Optional: show only new issues if it's a pull request. The default value is `false`.
           # only-new-issues: true
@@ -44,6 +44,3 @@ jobs:
 
           # Optional: if set to true then the action don't cache or restore ~/.cache/go-build.
           # skip-build-cache: true
-run:
-  skip-dirs:
-    - ibc/
