+++ .github/workflows/gosec.yml
@@ -17,4 +17,4 @@ jobs:
       - name: Run Gosec Security Scanner
         uses: securego/gosec@master
         with:
-          args: -exclude-dir=ibc ./... -exclude-generated ./...
+          args: -exclude-dir=legacy_ibc_testing ./... -exclude-generated ./...
