+++ app/consumer-democracy/proposals_whitelisting_test.go
@@ -4,7 +4,7 @@ import (
 	"testing"
 
 	appConsumer "github.com/cosmos/interchain-security/app/consumer-democracy"
-	ibctesting "github.com/cosmos/interchain-security/ibc/testing"
+	ibctesting "github.com/cosmos/interchain-security/legacy_ibc_testing/testing"
 	icstestingutils "github.com/cosmos/interchain-security/testutil/ibc_testing"
 	"github.com/stretchr/testify/require"
 )
