+++ sonar-project.properties
@@ -14,7 +14,7 @@ sonar.organization=cosmos
 # All golang artifacts
 sonar.sources=.
 # Do not calculate coverage metrics for statements in these files
-sonar.exclusions=**/vendor/**,**/*.pb.go,**/*.pb.gw.go,proto,**/*_test.go,tests/**,testutil/**,ibc/**,app/consumer/app.go
+sonar.exclusions=**/vendor/**,**/*.pb.go,**/*.pb.gw.go,proto,**/*_test.go,tests/**,testutil/**,legacy_ibc_testing/**,app/consumer/app.go
 sonar.tests=.
 # Run unit and e2e tests, but not integration tests
 sonar.test.inclusions=**/*_test.go
