+++ testutil/simibc/relay_util.go
@@ -7,8 +7,8 @@ import (
 	channeltypes "github.com/cosmos/ibc-go/v3/modules/core/04-channel/types"
 	host "github.com/cosmos/ibc-go/v3/modules/core/24-host"
 	ibctmtypes "github.com/cosmos/ibc-go/v3/modules/light-clients/07-tendermint/types"
-	simapp "github.com/cosmos/interchain-security/ibc/simapp"
-	ibctesting "github.com/cosmos/interchain-security/ibc/testing"
+	simapp "github.com/cosmos/interchain-security/legacy_ibc_testing/simapp"
+	ibctesting "github.com/cosmos/interchain-security/legacy_ibc_testing/testing"
 	"github.com/stretchr/testify/require"
 	tmtypes "github.com/tendermint/tendermint/types"
 )
